+++ scrollytelling/src/context/index.tsx
@@ -1,0 +1,41 @@
+import * as React from "react";
+
+export type ScrollytellingContextType = {
+  timeline: gsap.core.Timeline | undefined;
+  rootRef: React.RefObject<HTMLDivElement>;
+};
+
+export const ScrollytellingContext = React.createContext<
+  undefined | ScrollytellingContextType
+>(undefined);
+
+export const useScrollytelling = () => {
+  const scrollytelling = React.useContext(ScrollytellingContext);
+  if (!scrollytelling) {
+    throw new Error(
+      "useScrollytelling must be used within a Scrollytelling.Root"
+    );
+  }
+  return scrollytelling;
+};
+
+export type ScrollytellingDispatchersContextType = {
+  getTimelineSpace: (params: { start: number; end: number }) => {
+    duration: number;
+    position: number;
+    cleanup: () => void;
+  };
+  scopedQuerySelector: gsap.utils.SelectorFunc | undefined;
+};
+
+export const ScrollytellingDispatchersContext = React.createContext<
+  undefined | ScrollytellingDispatchersContextType
+>(undefined);
+
+export const useDispatcher = () => {
+  const dispatcher = React.useContext(ScrollytellingDispatchersContext);
+  if (!dispatcher) {
+    throw new Error("useDispatcher must be used within a ScrollytellingRoot");
+  }
+  return dispatcher;
+};
